pyproject.toml

[tool.mypy]
strict_optional = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts= """
    --html=test_results/results.html \
    --self-contained-html \
    --junitxml=test_results/results.xml \
    --disable-warnings
"""
log_file_level="INFO"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F", "PL", "YTT", "I"]
ignore = ["E501", "PLR2004"]

fixable = ["I"]

[tool.ruff.pylint]
max-args = 10

[tool.isort]
profile = "black"

[tool.setuptools_scm]

[tool.poetry]
name = "rubrical"
version = "0.0.0"
description = "A CLI to encourage (😅) people to update their dependencies!"
authors = ["Ivan Lee <ivanklee86@gmail.com>"]
license = "MPL-2.0"
homepage = "https://github.com/ivanklee86/rubrical"
repository = "https://github.com/ivanklee86/rubrical"
readme = "README.md"
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
]

[tool.poetry.scripts]
rubrical = "rubrical.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.7.0"}
pydantic = "^1.10.4"
semver = "^2.13.0"
pygithub = "^1.58.0"
python-benedict = {extras = ["all"], version = "^0.30.0"}
pyyaml = "^6.0"
requirements-parser = "^0.5.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
mypy = "^1.0"
pytest = "^7.2.0"
pylint = "^2.15.7"
pytest-cov = "^4.0.0"
pytest-flake8 = "^1.1.1"
pytest-html = "^3.2.0"
pytest-mock = "^3.10.0"
pytest-xdist = "^3.0.2"
setuptools-scm = "^7.0.5"
twine = "^4.0.2"
pre-commit = "^3.0.0"
isort = "^5.10.1"
ruff = "^0.0.260"
python-dotenv = "^1.0.0"
typer-cli = "^0.0.13"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
